moved toggle_parameter to drone_management.py
