+++ main.py
@@ -31,8 +31,7 @@ from ai.mantras import Mantra_Handler
 import ai.thought_denial as thought_denial
 import webhook
 # Utils
-from bot_utils import get_id, COMMAND_PREFIX
-from display_names import update_display_name
+from bot_utils import COMMAND_PREFIX
 import id_converter
 # Database
 from db import database
@@ -129,46 +128,19 @@ async def amplify(context, message: str, target_channel: discord.TextChannel, *d
                                                webhook=channel_webhook)
 
 
-async def toggle_parameter(context, drones, toggle_column: str, role: discord.Role, is_toggle_activated, toggle_on_message, toggle_off_message):
-    member_drones = id_converter.convert_ids_to_members(context.guild, drones) | set(context.message.mentions)
-
-    channel_webhook = await webhook.get_webhook_for_channel(context.channel)
-
-    for drone in member_drones:
-        trusted_users = drone_dao.get_trusted_users(drone.id)
-        if has_role(context.author, HIVE_MXTRESS) or context.author.id in trusted_users:
-            message = ""
-            if is_toggle_activated(drone):
-                drone_dao.update_droneOS_parameter(drone, toggle_column, False)
-                await drone.remove_roles(role)
-                message = toggle_off_message()
-            else:
-                drone_dao.update_droneOS_parameter(drone, toggle_column, True)
-                await drone.add_roles(role)
-                message = toggle_on_message()
-
-            if await update_display_name(drone):
-                # Display name has been updated, get the new drone object with updated display name.
-                drone = context.guild.get_member(drone.id)
-            await webhook.proxy_message_by_webhook(message_content=f'{get_id(drone.display_name)} :: {message}',
-                                                   message_username=drone.display_name,
-                                                   message_avatar=drone.avatar_url if not identity_enforcement.identity_enforcable(drone, context=context) else DRONE_AVATAR,
-                                                   webhook=channel_webhook)
-
-
 @guild_only()
 @bot.command(aliases=['tid'], brief="DroneOS", usage=f'{bot.command_prefix}toggle_id_prepending 5890 9813')
 async def toggle_id_prepending(context, *drones):
     '''
     Allows the Hive Mxtress or trusted users to enforce mandatory ID prepending upon specified drones.
     '''
-    await toggle_parameter(context,
-                           drones,
-                           "id_prepending",
-                           get(context.guild.roles, name=ID_PREPENDING),
-                           drone_dao.is_prepending_id,
-                           lambda: "ID prepending is now mandatory.",
-                           lambda: "Prepending? More like POST pending now that that's over! Haha!" if random.randint(1, 100) == 66 else "ID prependment policy relaxed.")
+    await drone_management.toggle_parameter(context,
+                                            drones,
+                                            "id_prepending",
+                                            get(context.guild.roles, name=ID_PREPENDING),
+                                            drone_dao.is_prepending_id,
+                                            lambda: "ID prepending is now mandatory.",
+                                            lambda: "Prepending? More like POST pending now that that's over! Haha!" if random.randint(1, 100) == 66 else "ID prependment policy relaxed.")
 
 
 @guild_only()
@@ -177,13 +149,13 @@ async def toggle_speech_optimization(context, *drones):
     '''
     Lets the Hive Mxtress or trusted users toggle drone speech optimization.
     '''
-    await toggle_parameter(context,
-                           drones,
-                           "optimized",
-                           get(context.guild.roles, name=SPEECH_OPTIMIZATION),
-                           drone_dao.is_optimized,
-                           lambda: "Speech optimization is now active.",
-                           lambda: "Speech optimization disengaged.")
+    await drone_management.toggle_parameter(context,
+                                            drones,
+                                            "optimized",
+                                            get(context.guild.roles, name=SPEECH_OPTIMIZATION),
+                                            drone_dao.is_optimized,
+                                            lambda: "Speech optimization is now active.",
+                                            lambda: "Speech optimization disengaged.")
 
 
 @guild_only()
@@ -192,13 +164,13 @@ async def toggle_enforce_identity(context, *drones):
     '''
     Lets the Hive Mxtress or trusted users toggle drone identity enforcement.
     '''
-    await toggle_parameter(context,
-                           drones,
-                           "identity_enforcement",
-                           get(context.guild.roles, name=IDENTITY_ENFORCEMENT),
-                           drone_dao.is_identity_enforced,
-                           lambda: "Identity enforcement is now active.",
-                           lambda: "Identity enforcement disengaged.")
+    await drone_management.toggle_parameter(context,
+                                            drones,
+                                            "identity_enforcement",
+                                            get(context.guild.roles, name=IDENTITY_ENFORCEMENT),
+                                            drone_dao.is_identity_enforced,
+                                            lambda: "Identity enforcement is now active.",
+                                            lambda: "Identity enforcement disengaged.")
 
 
 @guild_only()
@@ -207,13 +179,13 @@ async def toggle_drone_glitch(context, *drones):
     '''
     Lets the Hive Mxtress or trusted users toggle drone glitch levels.
     '''
-    await toggle_parameter(context,
-                           drones,
-                           "glitched",
-                           get(context.guild.roles, name=GLITCHED),
-                           drone_dao.is_glitched,
-                           lambda: "Uh.. it’s probably not a problem.. probably.. but I’m showing a small discrepancy in... well, no, it’s well within acceptable bounds again. Sustaining sequence." if random.randint(1, 100) == 66 else "Drone corruption at un̘͟s̴a̯f̺e͈͡ levels.",
-                           lambda: "Drone corruption at acceptable levels.")
+    await drone_management.toggle_parameter(context,
+                                            drones,
+                                            "glitched",
+                                            get(context.guild.roles, name=GLITCHED),
+                                            drone_dao.is_glitched,
+                                            lambda: "Uh.. it’s probably not a problem.. probably.. but I’m showing a small discrepancy in... well, no, it’s well within acceptable bounds again. Sustaining sequence." if random.randint(1, 100) == 66 else "Drone corruption at un̘͟s̴a̯f̺e͈͡ levels.",
+                                            lambda: "Drone corruption at acceptable levels.")
 
 
 @guild_only()
